libstdc++: PR 81857 std::copy_n increments iterators n times
Enhance std::istream_iterator implementation to avoid reading an element
on instantiation or on incrementation. Read is only perform when we use
the iterator to check if we reach the end of stream or if actually get the
read value.

Thanks to this change std::copy_n can be implemented to increment both input
and output iterators n times.

	PR libstdc++/81857
	* include/bits/stl_algo.h (__copy_n_a): Increment input and
	output iterators exactly n times.
	* include/bits/stream_iterator.h (istream_iterator::_M_stream): Mutable.
	(istream_iterator::_M_value): Mutable.
	(istream_iterator::_M_value): Mutable.
	(istream_iterator::_M_value_read): New, mutable.
	(istream_iterator(istream_type&)): Do not call _M_read.
	* testsuite/25_algorithms/copy_n/50119.cc (test01): Add check when
	calling copy_n of 0 element.
	* testsuite/25_algorithms/copy_n/81857.cc: New.