libstdc++ Replace std::deque iterator _M_last
In order to let the compiler know that the std::deque iterator _M_last is just
an offset after _M_first we introduce _M_last(). _M_last is just kept for abi
backward compatibility but rename in _M_last_ptr.
It is removed in _GLIBCXX_INLINE_VERSION.

libstdc++-v3/ChangeLog:

	* include/bits/stl_deque.h (_Deque_iterator<>::_S_buffer_size): Add C++11
	constexpr qualification.
	(_Deque_iterator<>::_M_last): Rename in...
	(_Deque_iterator<>::_M_last_ptr): ...that.
	[_GLIBCXX_INLINE_VERSION](_Deque_iterator<>::_M_last_ptr): Remove.
	(_Deque_iterator<>::_M_last()): New.
	(deque<>::_S_buffer_size): Add C++11 constexpr qualification, call iterator
	_S_buffer_size.
	* include/bits/deque.tcc: Adapt.
	* python/libstdcxx/v6/printers.py (StdDequePrinter): Adapt.