libstdc++: [_Hashtable] Enhance detection of end-of-bucket
To find out if we are still in the same bucket we can first check that current node
is not the next bucket's before-begin, the user can managed to build the container
this way. Then check if hash code are equals when cached.
At last resort to compute node bucket index through a modulo operation.

libstdc++-v3/ChangeLog:

	* include/bits/hashtable_policy.h (_Hashtable_base<>::_S_cached_hash_code_equals): New.
	* include/bits/hashtable.h (_Hashtable<>::_M_is_in_bucket): New, use latter.
	(_Hashtable<>::_M_find_before_node): Use latter.
	(_Hashtable<>::_M_find_before_node_tr): Likewise.
	* include/bits/hashtable_policy.h (_Equality<>::_M_equal): Likewise.