libstdc++: [_Hashtable] Avoid temporaries when inserting existing key
Following PR 115285 fix, the hashtable _S_forward_key responsible for finding
the best way to transmit the key part of the argument parameter to the hash
functor was modified to return a key_type instance for any argument type. Only
the overloads for const key_type& and key_type&& were still preventing a key_type
instanciation.

It revealed that a overload for key_type& was missing. To fix this problem we now
deal with all sort of key_type of reference type in the unique _S_forward_key static
method and remove the overloads. When decayed type of the input instance is the same
as key_type we avoid the instanciation. Note that for the comparison we also decay
key_type to properly managed when unordered container is instantiated with a const
type.

libstdc++-v3/ChangeLog:

	* include/bits/hashtable_policy.h
	(_ConvertToValueType<_Select1st, _Value>::operator(std::pair<>&)): New,
	allow consistent behavior with _ConvertToValueType<_Identity, _Value>.
	* include/bits/hashtable.h
	(_S_forward_key<_Kt>(_Kt&&)): Adapt to return _Kt&& when decayed _Kt is
	key_type, return a new key_type instance otherwise.
	(_S_forward_key(const key_type&)): Remove.
	(_S_forward_key(key_type&&)): Remove.
	* testsuite/23_containers/unordered_set/96088.cc (test03): Adapt expected
	allocation increment.